maded Sector::objectCount unsigned again as double-frees are fixed now

+++ trunk/src/common/bgloader/loader.h
@@ -398,7 +398,7 @@ private:
         bool checked;
         csString culler;
         csColor ambient;
-        int objectCount;
+        size_t objectCount;
         csRef<iSector> object;
         csWeakRef<Zone> parent;
         csRefArray<MeshGen> meshgen;
